upload_pipeline/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
urllib3>=2.0.0
selenium>=4.15.0
internetarchive>=3.5.0
